fix: Parse multiline array brackets correctly
Before, if array brackets were on different lines with whitespace and
generally uneven (though still valid TOML), they would not be parsed
correctly and would result in the wrong data output or an error that the
array was not completed or terminated correctly.

Now, using regex, it searches for the amount of opening brackets until
the first alpha-numeric character, and searches the string at the end
for brackets in the same way. The search continues until the matched
brackets at the end are the same size as the matched brackets at the
beginning. If there are not enough brackets to close the array properly,
the correct error is thrown.

Resolves: #9