fix: Properly escape characters on encode/decode
The lack of correct parsing on decode was due to insufficient regex. The
lack of escaping on the encoding was due to not having any string
replacement to apply the escape to the common characters. The extra line
at the end of encoded output was because the final result wasn't trimmed
for whitespace.

I have deleteme.TOML file as an input with the following contents:
```
a = "A\\ \" B"
b = "A\nB"
c = "this \t is my \u0610 last \\ big \\resort"
```
Now I run the following commands and get the following outputs:
```
>> a = toml.read('deleteme.toml')
a =
  struct with fields:

    a: 'A\ " B'
    b: 'A↵B'
    c: 'this → is my ؐ last \ big \resort'
>> toml.encode(a)
ans =
    'a = "A\\ \" B"
     b = "A\nB"
     c = "this \t is my \u0610 last \\ big \\resort"'
```

If I test nested keys: The input TOML has the following:
```
a = "A\\ \" B"
b.a = "A\nB"
b.b = "Nested \t Test \U0400"
c = "this \t is my \u0610 last \\ big \\resort"
```
The round-trip in MATLAB is the following:
```
>> a = toml.read('deleteme.toml')
a =
  struct with fields:

    a: 'A\ " B'
    b: [1×1 struct]
    c: 'this → is my ؐ last \ big \resort'
>> a.b
ans =
  struct with fields:

    a: 'A↵B'
    b: 'Nested → Test Ѐ'
>> toml.encode(a)
ans =
    'a = "A\\ \" B"
     c = "this \t is my \u0610 last \\ big \\resort"
     [b]
     a = "A\nB"
     b = "Nested \t Test Ѐ"'
```

Resolves: #4