fix: Sort structs correctly during encoding
Previously, multinested structs passed in for encoding were not being
sorted and ordered correctly. The results would cause the encoded TOML
output to have a key(field) under a different table(struct) than was
intended by the MATLAB instance.

Proper sorting happened only on the first pass of encoding which is why
it would always fail on nested structs. I moved the sorting routine to
its own function which is now called during the struct encoding portion
in repr.m. The sorting has been removed from the encoding main function
as it would be redundant.

Resolves: #3