fix: Encode numeric arrays correctly up to 3D
This partially fixes the encoding of numerical arrays. Before a
multidimensional array would be flatted to a single row. Now, up to 3
dimensions, it will correctly wrap the values in brackets. For single
row arrays, it adds an extra set of brackets. This isn't great but it
doesn't blow anything up on a round trip encode/decode. It'd be nice if
there were a clever way to fix it. Same with dimensions over 3. It's a
tricky problem to keep track of all the dimensions in a robust way.

Resolves: #11